refactor: Extension 파일 구분 및 불필요한 주석 제거
