+++ BoxOffice/Source/Extension/UIImage+.swift
@@ -1,0 +1,23 @@
+//
+//  UIImage+.swift
+//  BoxOffice
+//
+//  Created by seohyeon park on 2023/01/06.
+//
+
+import UIKit
+
+extension UIImage {
+    func resize(newWidth: CGFloat) -> UIImage {
+        let scale = newWidth / self.size.width
+        let newHeight = self.size.height * scale
+
+        let size = CGSize(width: newWidth, height: newHeight)
+        let render = UIGraphicsImageRenderer(size: size)
+        let renderImage = render.image { context in
+            self.draw(in: CGRect(origin: .zero, size: size))
+        }
+
+        return renderImage
+    }
+}
