+++ BoxOffice/Source/Model/DailyBoxOffice.swift
@@ -6,7 +6,7 @@
 //
 
 import Foundation
-//TODO: 파일 분리 및 폴더 정리 필요
+
 //MARK: 일일 박스오피스 DTO
 struct BoxOfficeResult: Decodable {
     let boxOfficeResult: DailyBoxOffice
@@ -19,11 +19,11 @@ struct DailyBoxOffice: Decodable {
 }
 
 struct BoxOfficeInfo: Decodable {
-    let rank: String // 순위
-    let rankInten: String // 순위 증감분
-    let rankOldAndNew: String // 순위 진입 여부
-    let movieNm: String // 영화 이름
-    let openDt: String // 개봉일
-    let audiAcc: String // 누적 관객
-    let movieCd: String // 영화 코드
+    let rank: String
+    let rankInten: String
+    let rankOldAndNew: String
+    let movieNm: String
+    let openDt: String
+    let audiAcc: String
+    let movieCd: String
 }
